fix: formlistaggregationresult type for ratingcount from integer to long to match enitity

+++ backend/src/main/java/format/backend/form/entity/FormListAggregationResult.java
@@ -32,7 +32,7 @@ public record FormListAggregationResult(
 
         @NonNull Long submissionsCount,
 
-        @NonNull Integer questionsCount,
+        @NonNull Long questionsCount,
 
         @NonNull Long ratingCount,
